requirements.txt
# Core Framework
streamlit==1.28.1

# LLM Providers
groq==0.4.2
openai==1.3.7
google-generativeai==0.3.0

# RAG & Embeddings (FIXED VERSIONS)
sentence-transformers==3.0.1
huggingface_hub==0.20.0
transformers==4.37.2
torch==2.1.2
numpy==1.24.3
scikit-learn==1.3.0

# Document Processing
PyPDF2==3.0.1

# Web Search
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
Pillow==10.0.0